+++ src/BusinessLogic/Tasks/GetDefaultParcelAndWarehouseTask.php
@@ -36,6 +36,21 @@ class GetDefaultParcelAndWarehouseTask extends Task
     }
 
     /**
+     * @inheritDoc
+     */
+    public function __serialize()
+    {
+        return $this->toArray();
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function __unserialize($data)
+    {
+    }
+
+    /**
      * Runs task logic.
      *
      * @throws \Logeecom\Infrastructure\Http\Exceptions\HttpAuthenticationException