+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## Unreleased changes
+### Added
+- Added __serialize() and __unserialize() methods to Serializable interface in order to add compatibility with PHP8.1.
+- Implemented __serialize() and __unserialize() methods in all classes that implement Serializable interface.
+
 ## [3.3.6](https://github.com/packlink-dev/ecommerce_module_core/compare/v3.3.5...v3.3.6) - 2022-02-09
 ### Changed
 - Updated marketing calls flag to always be set to true when registering a new client through modules.